+++ src/Drivers/MFUpdate/Storage/BlockStorageUpdate.cpp
@@ -83,7 +83,7 @@ BOOL BlockStorageUpdate::InitializeFiles( UINT32 blockTypes )
                 }
             }
         }
-        else if((unsigned long) freeListIdx < ARRAYSIZE(g_BlockStorageUpdate.m_freeList))
+        else if((unsigned int) freeListIdx < ARRAYSIZE(g_BlockStorageUpdate.m_freeList))
         {
 
             int size = g_BlockStorageUpdate.m_freeList[freeListIdx].File.Size;
@@ -117,7 +117,7 @@ INT32 BlockStorageUpdate::GetFreeHandle()
 {
     INT32 newHandle = -1;
 
-    for(int idx = 0; (unsigned long) idx < ARRAYSIZE(g_BlockStorageUpdate.m_files); idx++)
+    for(int idx = 0; (unsigned int) idx < ARRAYSIZE(g_BlockStorageUpdate.m_files); idx++)
     {
         if(g_BlockStorageUpdate.m_files[idx].Size == 0)
         {
@@ -158,7 +158,7 @@ INT32 BlockStorageUpdate::Create( MFUpdateHeader& storageHeader, UINT32 flags )
     {
         if( g_BlockStorageUpdate.m_files[i].Type    == storageHeader.UpdateType    && 
             g_BlockStorageUpdate.m_files[i].SubType == storageHeader.UpdateSubType && 
-            g_BlockStorageUpdate.m_files[i].ID      == (unsigned int) storageHeader.UpdateID      && 
+            g_BlockStorageUpdate.m_files[i].ID      == storageHeader.UpdateID      && 
             g_BlockStorageUpdate.m_files[i].Size    != 0 )
         {
             return -1;
@@ -212,7 +212,7 @@ INT32 BlockStorageUpdate::Create( MFUpdateHeader& storageHeader, UINT32 flags )
     return newHandle;
 }
 
-INT32 BlockStorageUpdate::Open( INT32 storageID, UINT16 storageType, UINT16 storageSubType)
+INT32 BlockStorageUpdate::Open( UINT32 storageID, UINT16 storageType, UINT16 storageSubType)
 {
     INT32 i;
     struct UpdateBlockHeader header;
@@ -223,7 +223,7 @@ INT32 BlockStorageUpdate::Open( INT32 storageID, UINT16 storageType, UINT16 stor
 
     for(i = 0; (unsigned int) i <  ARRAYSIZE(g_BlockStorageUpdate.m_files); i++)
     {
-        if( (MFUPDATE_UPDATEID_ANY      == (unsigned int) storageID      || g_BlockStorageUpdate.m_files[i].ID      == (unsigned int) storageID     ) && 
+        if( (MFUPDATE_UPDATEID_ANY      == storageID      || g_BlockStorageUpdate.m_files[i].ID      == storageID     ) && 
             (MFUPDATE_UPDATETYPE_ANY    == storageType    || g_BlockStorageUpdate.m_files[i].Type    == storageType   ) &&
             (MFUPDATE_UPDATESUBTYPE_ANY == storageSubType || g_BlockStorageUpdate.m_files[i].SubType == storageSubType) &&
             g_BlockStorageUpdate.m_files[i].Size > 0)
@@ -241,11 +241,11 @@ void  BlockStorageUpdate::Close     ( INT32 handleStorage )
     (void) handleStorage;
 }
 
-BOOL BlockStorageUpdate::Delete( INT32 storageID, UINT16 storageType, UINT16 storageSubType )
+BOOL BlockStorageUpdate::Delete( UINT32 storageID, UINT16 storageType, UINT16 storageSubType )
 {
     for(int i = 0; (unsigned int) i < ARRAYSIZE(g_BlockStorageUpdate.m_files); i++)
     {
-        if( (MFUPDATE_UPDATEID_ANY      == (unsigned int) storageID   || g_BlockStorageUpdate.m_files[i].ID      == (unsigned int) storageID     ) && 
+        if( (MFUPDATE_UPDATEID_ANY      == storageID   || g_BlockStorageUpdate.m_files[i].ID      == storageID     ) && 
             (MFUPDATE_UPDATETYPE_ANY    == storageType || g_BlockStorageUpdate.m_files[i].Type    == storageType   ) &&
             (MFUPDATE_UPDATESUBTYPE_ANY == storageType || g_BlockStorageUpdate.m_files[i].SubType == storageSubType) &&
             g_BlockStorageUpdate.m_files[i].Size > 0)
@@ -270,7 +270,7 @@ BOOL BlockStorageUpdate::Delete( INT32 storageID, UINT16 storageType, UINT16 sto
                     }
                     else
                     {
-                        for(int j = 0; (unsigned long) j < ARRAYSIZE(g_BlockStorageUpdate.m_freeList); j++)
+                        for(int j = 0; (unsigned int) j < ARRAYSIZE(g_BlockStorageUpdate.m_freeList); j++)
                         {
                             if(g_BlockStorageUpdate.m_freeList[j].File.Size == 0)
                             {
@@ -343,7 +343,7 @@ BOOL BlockStorageUpdate::GetFiles( UINT16 storageType, INT32* storageIDs, INT32*
 
 BOOL BlockStorageUpdate::IsErased ( INT32  handleStorage, INT32 fileOffset, INT32  len )
 {
-    if(handleStorage < 0 || (unsigned int) handleStorage >= ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
+    if(handleStorage < 0 || handleStorage >= (signed int) ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
 
     BlockStorageStream_Seek(&g_BlockStorageUpdate.m_stream, g_BlockStorageUpdate.m_files[handleStorage].StartAddress + sizeof(UpdateBlockHeader) + fileOffset, SeekOrigin::BlockStorageStream_SeekBegin);
         
@@ -352,7 +352,7 @@ BOOL BlockStorageUpdate::IsErased ( INT32  handleStorage, INT32 fileOffset, INT3
 
 INT32 BlockStorageUpdate::Write( INT32  handleStorage, INT32 fileOffset, UINT8* pData, INT32 len )
 {
-    if(handleStorage < 0 || (unsigned int) handleStorage >= ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
+    if(handleStorage < 0 || handleStorage >= (signed int) ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
 
     BlockStorageStream_Seek(&g_BlockStorageUpdate.m_stream, g_BlockStorageUpdate.m_files[handleStorage].StartAddress + sizeof(UpdateBlockHeader) + fileOffset, SeekOrigin::BlockStorageStream_SeekBegin );
         
@@ -361,7 +361,7 @@ INT32 BlockStorageUpdate::Write( INT32  handleStorage, INT32 fileOffset, UINT8* 
 
 INT32 BlockStorageUpdate::Read( INT32  handleStorage, INT32 fileOffset, UINT8* pData, INT32 len )
 {
-    if(handleStorage < 0 || (unsigned int) handleStorage >= ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
+    if(handleStorage < 0 || handleStorage >= (signed int) ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
 
     BlockStorageStream_Seek(&g_BlockStorageUpdate.m_stream, g_BlockStorageUpdate.m_files[handleStorage].StartAddress + sizeof(UpdateBlockHeader) + fileOffset, SeekOrigin::BlockStorageStream_SeekBegin );
         
@@ -372,7 +372,7 @@ BOOL BlockStorageUpdate::GetHeader( INT32 handleStorage, MFUpdateHeader* pHeader
 {
     UpdateBlockHeader hdr;
     
-    if(handleStorage < 0 || (unsigned int) handleStorage >= ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
+    if(handleStorage < 0 || handleStorage >= (signed int) ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
 
     BlockStorageStream_Seek(&g_BlockStorageUpdate.m_stream, g_BlockStorageUpdate.m_files[handleStorage].StartAddress, SeekOrigin::BlockStorageStream_SeekBegin );
         
@@ -385,7 +385,7 @@ BOOL BlockStorageUpdate::GetHeader( INT32 handleStorage, MFUpdateHeader* pHeader
 
 UINT32 BlockStorageUpdate::GetEraseSize( INT32 handleStorage )
 {
-    if(handleStorage < 0 || (unsigned int) handleStorage >= ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
+    if(handleStorage < 0 || handleStorage >= (signed int) ARRAYSIZE(g_BlockStorageUpdate.m_files) || g_BlockStorageUpdate.m_files[handleStorage].Size <= 0) return FALSE;
 
     return g_BlockStorageUpdate.m_stream.BlockLength;
 }
